refactor: renamed pascal-cased and camel-cased file to skewer-cased for consistency
