+++ src/interface-checker.ts
